requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.115.6
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.2
idna==3.11
Jinja2==3.1.4
jiter==0.13.0
loguru==0.7.2
MarkupSafe==3.0.3
multidict==6.7.1
openai==1.45.0
propcache==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.11.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.3
requests==2.32.5
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.41.3
tqdm==4.67.3
twilio==9.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.30.6
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
